+++ dev-packages/e2e-tests/test-applications/nextjs-14/app/propagation/test-outgoing-http/route.ts
@@ -1,0 +1,9 @@
+import { NextResponse } from 'next/server';
+import { makeHttpRequest } from '../utils';
+
+export const dynamic = 'force-dynamic';
+
+export async function GET() {
+  const data = await makeHttpRequest(`http://localhost:3030/propagation/test-outgoing-http/check`);
+  return NextResponse.json(data);
+}