+++ packages/node/src/integrations/tracing/express.ts
@@ -1,8 +1,8 @@
 import type * as http from 'http';
-import { registerInstrumentations } from '@opentelemetry/instrumentation';
 import { ExpressInstrumentation } from '@opentelemetry/instrumentation-express';
 import { defineIntegration, getDefaultIsolationScope } from '@sentry/core';
 import { captureException, getClient, getIsolationScope } from '@sentry/core';
+import { addOpenTelemetryInstrumentation } from '@sentry/opentelemetry';
 import type { IntegrationFn } from '@sentry/types';
 
 import { logger } from '@sentry/utils';
@@ -14,29 +14,27 @@ const _expressIntegration = (() => {
   return {
     name: 'Express',
     setupOnce() {
-      registerInstrumentations({
-        instrumentations: [
-          new ExpressInstrumentation({
-            requestHook(span) {
-              addOriginToSpan(span, 'auto.http.otel.express');
-            },
-            spanNameHook(info, defaultName) {
-              if (getIsolationScope() === getDefaultIsolationScope()) {
-                DEBUG_BUILD &&
-                  logger.warn('Isolation scope is still default isolation scope - skipping setting transactionName');
-                return defaultName;
-              }
-              if (info.layerType === 'request_handler') {
-                // type cast b/c Otel unfortunately types info.request as any :(
-                const req = info.request as { method?: string };
-                const method = req.method ? req.method.toUpperCase() : 'GET';
-                getIsolationScope().setTransactionName(`${method} ${info.route}`);
-              }
+      addOpenTelemetryInstrumentation(
+        new ExpressInstrumentation({
+          requestHook(span) {
+            addOriginToSpan(span, 'auto.http.otel.express');
+          },
+          spanNameHook(info, defaultName) {
+            if (getIsolationScope() === getDefaultIsolationScope()) {
+              DEBUG_BUILD &&
+                logger.warn('Isolation scope is still default isolation scope - skipping setting transactionName');
               return defaultName;
-            },
-          }),
-        ],
-      });
+            }
+            if (info.layerType === 'request_handler') {
+              // type cast b/c Otel unfortunately types info.request as any :(
+              const req = info.request as { method?: string };
+              const method = req.method ? req.method.toUpperCase() : 'GET';
+              getIsolationScope().setTransactionName(`${method} ${info.route}`);
+            }
+            return defaultName;
+          },
+        }),
+      );
     },
   };
 }) satisfies IntegrationFn;