+++ packages/node/src/integrations/tracing/koa.ts
@@ -1,4 +1,3 @@
-import { registerInstrumentations } from '@opentelemetry/instrumentation';
 import { KoaInstrumentation } from '@opentelemetry/instrumentation-koa';
 import { SEMATTRS_HTTP_ROUTE } from '@opentelemetry/semantic-conventions';
 import {
@@ -8,6 +7,7 @@ import {
   getIsolationScope,
   spanToJSON,
 } from '@sentry/core';
+import { addOpenTelemetryInstrumentation } from '@sentry/opentelemetry';
 import type { IntegrationFn } from '@sentry/types';
 import { logger } from '@sentry/utils';
 import { DEBUG_BUILD } from '../../debug-build';
@@ -16,31 +16,30 @@ const _koaIntegration = (() => {
   return {
     name: 'Koa',
     setupOnce() {
-      registerInstrumentations({
-        instrumentations: [
-          new KoaInstrumentation({
-            requestHook(span, info) {
-              if (getIsolationScope() === getDefaultIsolationScope()) {
-                DEBUG_BUILD &&
-                  logger.warn('Isolation scope is default isolation scope - skipping setting transactionName');
-                return;
-              }
-              const attributes = spanToJSON(span).data;
-              const route = attributes && attributes[SEMATTRS_HTTP_ROUTE];
-              const method = info.context.request.method.toUpperCase() || 'GET';
-              if (route) {
-                getIsolationScope().setTransactionName(`${method} ${route}`);
-              }
-            },
-          }),
-        ],
-      });
+      addOpenTelemetryInstrumentation(
+        new KoaInstrumentation({
+          requestHook(span, info) {
+            if (getIsolationScope() === getDefaultIsolationScope()) {
+              DEBUG_BUILD &&
+                logger.warn('Isolation scope is default isolation scope - skipping setting transactionName');
+              return;
+            }
+            const attributes = spanToJSON(span).data;
+            const route = attributes && attributes[SEMATTRS_HTTP_ROUTE];
+            const method = info.context.request.method.toUpperCase() || 'GET';
+            if (route) {
+              getIsolationScope().setTransactionName(`${method} ${route}`);
+            }
+          },
+        }),
+      );
     },
   };
 }) satisfies IntegrationFn;
 
 export const koaIntegration = defineIntegration(_koaIntegration);
 
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const setupKoaErrorHandler = (app: { use: (arg0: (ctx: any, next: any) => Promise<void>) => void }): void => {
   app.use(async (ctx, next) => {
     try {