+++ packages/node/src/integrations/tracing/mongo.ts
@@ -1,6 +1,6 @@
-import { registerInstrumentations } from '@opentelemetry/instrumentation';
 import { MongoDBInstrumentation } from '@opentelemetry/instrumentation-mongodb';
 import { defineIntegration } from '@sentry/core';
+import { addOpenTelemetryInstrumentation } from '@sentry/opentelemetry';
 import type { IntegrationFn } from '@sentry/types';
 
 import { addOriginToSpan } from '../../utils/addOriginToSpan';
@@ -9,15 +9,13 @@ const _mongoIntegration = (() => {
   return {
     name: 'Mongo',
     setupOnce() {
-      registerInstrumentations({
-        instrumentations: [
-          new MongoDBInstrumentation({
-            responseHook(span) {
-              addOriginToSpan(span, 'auto.db.otel.mongo');
-            },
-          }),
-        ],
-      });
+      addOpenTelemetryInstrumentation(
+        new MongoDBInstrumentation({
+          responseHook(span) {
+            addOriginToSpan(span, 'auto.db.otel.mongo');
+          },
+        }),
+      );
     },
   };
 }) satisfies IntegrationFn;