+++ packages/sveltekit/src/server/index.ts
@@ -71,6 +71,7 @@ export {
   trpcMiddleware,
   spanToJSON,
   spanToTraceHeader,
+  addOpenTelemetryInstrumentation,
 } from '@sentry/node';
 
 // We can still leave this for the carrier init and type exports